Moved resizeImage into utils.
