+++ src/js/file-manager.js
@@ -163,50 +163,4 @@ fileManager.isTooLarge = () => { return false; };
  */
 fileManager.getMaxSizeReadable = () => { return `${5}MB`; };
 
-/**
- * @function resizeImage
- *
- * @param {File} file - image file to be resized
- * @param {number} maxWidth - maximum width of resized image
- * @param {number} maxHeight - maximum height of resized image
- *
- * @return {Promise<Blob>} promise of resized image blob
- */
-fileManager.resizeImage = ( file, maxWidth, maxHeight ) => {
-    return new Promise( ( resolve, reject ) => {
-        let image = new Image();
-        image.src = URL.createObjectURL( file );
-        image.onload = () => {
-            let width = image.width;
-            let height = image.height;
-
-            if ( width <= maxWidth && height <= maxHeight ) {
-                resolve( file );
-            }
-
-            let newWidth;
-            let newHeight;
-
-            if ( width > height ) {
-                newHeight = height * ( maxWidth / width );
-                newWidth = maxWidth;
-            } else {
-                newWidth = width * ( maxHeight / height );
-                newHeight = maxHeight;
-            }
-
-            let canvas = document.createElement( 'canvas' );
-            canvas.width = newWidth;
-            canvas.height = newHeight;
-
-            let context = canvas.getContext( '2d' );
-
-            context.drawImage( image, 0, 0, newWidth, newHeight );
-
-            canvas.toBlob( resolve, file.type );
-        };
-        image.onerror = reject;
-    } );
-};
-
 export default fileManager;
